+++ plugin/controllers/models/servicelist.py
@@ -1,61 +1,54 @@
-##############################################################################
-#                        2011-2022 E2OpenPlugins                                  #
-#                                                                            #
-#  This file is open source software; you can redistribute it and/or modify  #
-#     it under the terms of the GNU General Public License version 2 as      #
-#               published by the Free Software Foundation.                   #
-#                                                                            #
-##############################################################################
+##########################################################################
+# OpenWebif: serviceslist
+##########################################################################
+# Copyright (C) 2011 - 2023 jbleyel and E2OpenPlugins
+#
+# This program is free software; you can redistribute it and/or modify it
+# under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software Foundation,
+# Inc., 51 Franklin Street, Fifth Floor, Boston MA 02110-1301, USA.
+##########################################################################
+
 from urllib.parse import unquote
 from enigma import eDVBDB
 from Components.NimManager import nimmanager
 import Components.ParentalControl
 
 
-def reloadLameDB(self):
-	self.eDVBDB.reloadServicelist()
-
-
-def reloadUserBouquets(self):
-	self.eDVBDB.reloadBouquets()
-
-
-def reloadTransponders(self):
-	nimmanager.readTransponders()
-
-
-def reloadParentalControl(self):
-	Components.ParentalControl.parentalControl.open()
-
-
-def reloadServicesLists(self, mode):
-	self.eDVBDB = eDVBDB.getInstance()
-	res = "False"
-	msg = "missing or wrong parameter mode [0=both, 1=lamedb only, 2=userbouqets only, 3=transponders, 4=parentalcontrol white-/blacklist]"
+def reloadServicesLists(mode):
+	msgs = []
 	mode = unquote(mode)
-	if mode == "0":
-		reloadLameDB(self)
-		reloadUserBouquets(self)
-		res = True
-		msg = "reloaded both"
-	elif mode == "1":
-		reloadLameDB(self)
-		res = True
-		msg = "reloaded lamedb"
-	elif mode == "2":
-		reloadUserBouquets(self)
-		res = True
-		msg = "reloaded bouquets"
-	elif mode == "3":
-		reloadTransponders(self)
-		res = True
-		msg = "reloaded transponders"
-	elif mode == "4":
-		reloadParentalControl(self)
-		res = True
-		msg = "reloaded parentalcontrol white-/blacklist"
+	modes = mode.split(",")
+	for mode in modes:
+		if mode in ("0", "2"):
+			eDVBDB.getInstance().reloadBouquets()
+			msgs.append("bouquets")
+		if mode in ("0", "1"):
+			eDVBDB.getInstance().reloadServicelist()
+			msgs.append("lamedb")
+		elif mode == "3":
+			nimmanager.readTransponders()
+			msgs.append("transponders")
+		elif mode == "4":
+			Components.ParentalControl.parentalControl.open()
+			msgs.append("parentalcontrol white-/blacklist")
 
-	return {
-		"result": res,
-		"message": msg
-	}
+	if msgs:
+		return {
+			"result": True,
+			"message": "reloaded %s" % " ".join(msgs)
+		}
+	else:
+		return {
+			"result": False,
+			"message": "missing or wrong parameter mode [0 = lamedb and userbouqets, 1 = lamedb only, 2 = userbouqets only, 3 = transponders, 4 = parentalcontrol white/blacklist"
+		}